refactor files according to the code review comments
