--- app/src/main/java/com/github/sdpcoachme/data/messaging/ReadState.kt
@@ -1,7 +1,0 @@
-package com.github.sdpcoachme.data.messaging
-
-enum class ReadState {
-        SENT,
-        RECEIVED,
-        READ,
-}
